+++ test/doc/algo/CMakeLists.txt
@@ -14,7 +14,7 @@ make_unit( "doc.algo" iota.cpp               )
 make_unit( "doc.algo" iota_with_step.cpp     )
 make_unit( "doc.algo" fill.cpp               )
 make_unit( "doc.algo" find.cpp               )
-make_unit( "doc.algo" find_backward.cpp      )
+make_unit( "doc.algo" find_last.cpp      )
 make_unit( "doc.algo" inclusive_scan.cpp     )
 make_unit( "doc.algo" map.cpp                )
 make_unit( "doc.algo" min.cpp                )
