+++ test/unit/algo/algorithm/find_last_if_not_generic.cpp
@@ -12,7 +12,7 @@
 
 #include <algorithm>
 
-template<typename TraitsSupport> struct not_not_find_last_if_ : TraitsSupport
+template<typename TraitsSupport> struct not_not_find_if_backward_ : TraitsSupport
 {
   template<typename Rng, typename P> EVE_FORCEINLINE auto operator()(Rng&& rng, P p) const
   {
@@ -21,14 +21,14 @@ template<typename TraitsSupport> struct not_not_find_last_if_ : TraitsSupport
   }
 };
 
-inline constexpr auto not_not_find_last_if =
-    eve::algo::function_with_traits<not_not_find_last_if_>[eve::algo::find_if_not.get_traits()];
+inline constexpr auto not_not_find_if_backward =
+    eve::algo::function_with_traits<not_not_find_if_backward_>[eve::algo::find_if_not.get_traits()];
 
-TTS_CASE_TPL("Check find_last_if", algo_test::selected_types)
+TTS_CASE_TPL("Check find_if_backward", algo_test::selected_types)
 <typename T>(tts::type<T>)
 {
   algo_test::find_generic_backward_test(eve::as<T> {},
-                                        not_not_find_last_if,
+                                        not_not_find_if_backward,
                                         [](auto f, auto l, auto expected, auto actual)
                                         {
                                           TTS_EQUAL(actual - f, expected - f);
