+++ test/unit/algo/algorithm/find_like_special_cases.cpp
@@ -53,7 +53,7 @@ TTS_CASE("eve.algo.find value")
   std::vector<int>::const_iterator found = eve::algo::find[eve::algo::no_aligning](v, 3);
   TTS_EQUAL((found - v.begin()), 2);
 
-  std::vector<int>::const_iterator found_b = eve::algo::find_backward[eve::algo::no_aligning](v, 3);
+  std::vector<int>::const_iterator found_b = eve::algo::find_last[eve::algo::no_aligning](v, 3);
   TTS_EQUAL(found_b - v.begin(), static_cast<int>(v.size() - 2));
 };
